new services

+++ backend/src/controllers/quizz.controller.ts
@@ -2,9 +2,9 @@ import { Request, Response, NextFunction } from "express";
 import { StatusCodes } from "http-status-codes";
 import Quizz from "../models/quizz.model";
 import { ObjectId } from "mongodb";
+import Challenge from "../models/challenge.model";
 
 class QuizzController {
-  
   static async GetSingleQuizz(req: Request, res: Response, next: NextFunction) {
     try {
       const quizId = req.params.id;
@@ -25,10 +25,9 @@ class QuizzController {
       next(error);
     }
   }
-
   static async AddQuizz(req: Request, res: Response, next: NextFunction) {
     try {
-      const { questions } = req.body;
+      const { questions, challengeId } = req.body;
       const user = req.user;
 
       const quiz = await Quizz.create({
@@ -36,9 +35,11 @@ class QuizzController {
         questions,
       });
 
+      const challenge = await Challenge.findByIdAndUpdate(challengeId, { $push: { quizzes: quiz._id } }, { new: true });
+
       return res.status(StatusCodes.OK).json({
         status: true,
-        id: quiz._id,
+        id: quiz._id
       });
     } catch (error) {
       next(error);
@@ -48,9 +49,8 @@ class QuizzController {
   static async EditQuizz(req: Request, res: Response, next: NextFunction) {
     try {
       const { quizzId, questions } = req.body;
-      console.log(quizzId, questions);
-      const quiz = await Quizz.findByIdAndUpdate(quizzId, { questions }, { new: true });
 
+      const quiz = await Quizz.findByIdAndUpdate(quizzId, { questions }, { new: true });
 
       if (!quiz) {
         return res.status(StatusCodes.NOT_FOUND).json({
@@ -70,7 +70,18 @@ class QuizzController {
 
   static async DeleteQuizz(req: Request, res: Response, next: NextFunction) {
     try {
-      const quizId = req.body;
+      const quizId = req.body.quizId;
+
+      // Find the corresponding challenge and remove the quiz's _id
+      const challenge = await Challenge.findOneAndUpdate({ quizzes: quizId }, { $pull: { quizzes: quizId } }, { new: true });
+
+      if (!challenge) {
+        return res.status(StatusCodes.NOT_FOUND).json({
+          status: false,
+          message: "Challenge not found for the given quiz",
+        });
+      }
+
       const quiz = await Quizz.findByIdAndDelete(quizId);
 
       if (!quiz) {
@@ -83,6 +94,7 @@ class QuizzController {
       return res.status(StatusCodes.OK).json({
         status: true,
         message: "Quizz deleted successfully",
+        challenge,
       });
     } catch (error) {
       next(error);
